Rename application errors to validation errors
Also add RESERVATION_TIME_CHANGE_NOT_ALLOWED error code

Refs TILA-2003
