+++ api/graphql/validation_errors.py
@@ -24,6 +24,7 @@ class ValidationErrorCodes(Enum):
     RESERVATION_NOT_WITHIN_ALLOWED_TIME_RANGE = (
         "RESERVATION_NOT_WITHIN_ALLOWED_TIME_RANGE"
     )
+    RESERVATION_MODIFICATION_NOT_ALLOWED = "RESERVATION_MODIFICATION_NOT_ALLOWED"
     RESERVATION_UNIT_TYPE_IS_SEASON = "RESERVATION_UNIT_TYPE_IS_SEASON"
     NO_PERMISSION = "NO_PERMISSION"
     CHANGES_NOT_ALLOWED = "CHANGES_NOT_ALLOWED"
